Multiple fixes
Rationalise the Bluetooth helper services so that they run before
bluetoothd, avoiding the need to restart it if the BT address changes.
At a later date the two services could possibly be unified.

These scripts should be revisited when the reason for the need to
power-cycle the interface to permit Secure Simple Pairing work is known.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>